Move components to UI.
