--- src/components/Dialog/index.js
@@ -1,4 +1,0 @@
-export { default } from './Dialog';
-export { default as Warning } from './Warning';
-export { default as Confirm } from './Confirm';
-export { default as Notice } from './Notice';
