+++ src/styles/_components.scss
@@ -8,7 +8,6 @@
 @import './components/stage-editor';
 @import './components/type-editor';
 @import './components/guided';
-@import './components/modal';
 @import './components/overview';
 @import './components/timeline';
 @import './components/code-view';
@@ -27,5 +26,8 @@
 @import './components/error';
 @import './components/issues';
 @import './components/list';
+<<<<<<< HEAD
 @import './components/attributes-table';
 @import './components/dialog';
+=======
+>>>>>>> 668e53c... Move components to UI.
