--- src/styles/components/_dialog.scss
@@ -1,14 +1,0 @@
-.dialog {
-  background: white;
-  display: flex;
-  flex-direction: column;
-
-  &__title {
-    background: grey;
-    padding: 20px;
-  }
-
-  &__content {
-    padding: 50px;
-  }
-}
